requirements.txt
streamlit
torch
open_clip_torch
faiss-cpu
pillow
pandas
rapidfuzz
transformers

